Egg Corridor beetles shouldn't act the same as Sand Zone beetles
For some reason, in this clone, the Egg Corridor beetles follow the Sand
Zone behaviour of only flying across when Quote is in their line-of-sight.
In the real Cave Story, Egg Corridor beetles fly across depending on a
timer, which, going by eye, looks to be a single second (at 60FPS). They
don't care where Quote is.